% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guerry_ranks.R
\docType{data}
\name{Guerry_ranks}
\alias{Guerry_ranks}
\title{Ranks of Data from A.-M. Guerry, "Essay on the Moral Statistics of France"}
\description{
This is the same data as in \code{\link{Guerry}}, except that all numeric
variables have been converted to ranks.
}
\seealso{
\code{\link{Angeville}} for other analysis variables
}
