\name{Guerry_ranks}
\Rdversion{1.1}
\alias{Guerry_ranks}
\docType{data}
\encoding{latin1}
\title{
Ranks of Data from A.-M. Guerry, "Essay on the Moral Statistics of France"
}
\description{
This is the same data as in \code{link{Guerry}}, except that all numberic variables have been converted to ranks.
}
\usage{data(Guerry_ranks)}
\seealso{
\code{\link{Angeville}} for other analysis variables
}

